requirements.txt
Pympler==0.7
orjson==1.35
ujson==2.0.11
guppy3==3.0.8
memory-profiler==0.55.0